+++ packages/markuplint/src/cli/bootstrap.ts
@@ -25,6 +25,7 @@ Options
 	--severity-parse-error                 Specifies the severity level of parse errors. Supports "error", "warning", and "off". Default: "error".
 	--max-count                            Limit the number of violations shown. Default: 0 (no limit).
 	--max-warnings                         Number of warnings to trigger nonzero exit code. Default: -1 (no limit).
+	--progressive-output                   Output results immediately after processing each file. Default: false.
 
 	--init                                 Initialize settings interactively.
 	--search                               Search lines of codes that include the target element by selectors.
@@ -118,6 +119,11 @@ export const cli = meow(help, {
 		maxWarnings: {
 			type: 'number',
 			default: -1,
+		},
+		progressiveOutput: {
+			type: 'boolean',
+			// TODO: It will be changed to `true` in the next major version.
+			default: false,
 		},
 	},
 });