+++ packages/markuplint/CHANGELOG.md
@@ -7,9 +7,11 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 **Note:** Version bump only for package markuplint
 
+# [Unreleased](https://github.com/markuplint/markuplint/compare/markuplint@4.13.0...HEAD)
 
+### Features
 
-
+- **markuplint:** add --progressive-output option for CLI ([9cbfa69](https://github.com/markuplint/markuplint/commit/9cbfa69d1))
 
 # [4.13.0](https://github.com/markuplint/markuplint/compare/markuplint@4.12.0...markuplint@4.13.0) (2025-08-13)